// src/_common/exceptions/error-code.enum.ts
export enum ErrorCodeEnum {
    UNAUTHORIZED = 600,
    INVALID_CREDENTIALS = 601,
    INVALID_USER = 602,
    NEW_PASSWORD_NOT_CONFIRMED = 603,
    OLD_PASSWORD_AND_NEW_ARE_MATCHED = 604,
    INCORRECT_PASSWORD = 605,
    USERNAME_OR_EMAIL_ALREADY_EXIST = 606,
    INVALID_TOKEN = 607,
    ALLOWED_FOR_ADMIN_ONLY = 608,
    BLOCKED_USER = 609,
    ACCOUNT_EXISTS = 610,
    PROVIDED_CREDENTIALS_ALREADY_EXISTS = 611,
    INVALID_PLATFORM = 612,
    UNKNOWN_ERROR = 613,
    INVALID_EMAIL = 614,
    ACCOUNT_EXISTS_WITH_DIFFERENT_PROVIDER_ID_ASK_TO_REPLACE = 615,
    ACCOUNT_EXISTS_ASK_TO_MERGE = 616,
    USER_DOES_NOT_EXIST = 617,
    INVALID_VERIFICATION_CODE = 618,
    USER_ALREADY_EXIST = 619,
    VERIFICATION_CODE_NOT_EXIST = 620,
    EXPIRED_VERIFICATION_CODE = 621,
}